+++ src/main/java/svenhjol/charmony/api/StoneCircleDefinitionsProvider.java
@@ -1,0 +1,8 @@
+package svenhjol.charmony.api;
+
+import java.util.List;
+
+@SuppressWarnings("unused")
+public interface StoneCircleDefinitionsProvider {
+    List<StoneCircleDefinition> getStoneCircleDefinitions();
+}